promql::extension_plan::range_manipulate

Struct RangeManipulate

pub struct RangeManipulate {
    start: Millisecond,
    end: Millisecond,
    interval: Millisecond,
    range: Millisecond,
    time_index: String,
    field_columns: Vec<String>,
    input: LogicalPlan,
    output_schema: DFSchemaRef,
}

Time series manipulator for range function.

This plan will โfoldโ time index and value columns into RangeArrays, and truncate other columns to the same length with the โfoldedโ RangeArray column.

To pass runtime information to the execution plan (or the range function), This plan will add those extra columns:

timestamp range with type RangeArray, which is the folded timestamp column.

end of current range with the same type as the timestamp column. (todo)

Fields
start: Millisecondend: Millisecondinterval: Millisecondrange: Millisecondtime_index: Stringfield_columns: Vec<String>input: LogicalPlanoutput_schema: DFSchemaRef
Implementations

impl RangeManipulate

pub fn new( start: Millisecond, end: Millisecond, interval: Millisecond, range: Millisecond, time_index: String, field_columns: Vec<String>, input: LogicalPlan, ) -> DataFusionResult<Self>

pub const fn name() -> &'static str

pub fn build_timestamp_range_name(time_index: &str) -> String

pub fn internal_range_end_col_name() -> String

fn range_timestamp_name(&self) -> String

fn calculate_output_schema( input_schema: &DFSchemaRef, time_index: &str, field_columns: &[String], ) -> DataFusionResult<DFSchemaRef>

pub fn to_execution_plan( &self, exec_input: Arc<dyn ExecutionPlan>, ) -> Arc<dyn ExecutionPlan>

pub fn serialize(&self) -> Vec<u8>

pub fn deserialize(bytes: &[u8]) -> Result<Self>