promql::extension_plan

Struct HistogramFoldExec

pub struct HistogramFoldExec {
    le_column_index: usize,
    input: Arc<dyn ExecutionPlan>,
    output_schema: SchemaRef,
    field_column_index: usize,
    ts_column_index: usize,
    quantile: f64,
    metric: ExecutionPlanMetricsSet,
    properties: PlanProperties,
}

Fields
le_column_index: usize

Index for le column in the schema of input.

input: Arc<dyn ExecutionPlan>output_schema: SchemaReffield_column_index: usize

Index for field column in the schema of input.

ts_column_index: usizequantile: f64metric: ExecutionPlanMetricsSetproperties: PlanProperties
Implementations

impl HistogramFoldExec

pub fn tag_col_exprs(&self) -> Vec<Arc<dyn PhysicalExpr>>

Return all the [PhysicalExpr] of tag columns in order.

Tag columns are all columns except le, field and ts columns.