Module promql::functions

Modules

aggr_over_time 🔒

changes 🔒

Implementation of changes in PromQL. Refer to the original implementation.

deriv 🔒

Implementation of deriv in PromQL. Refer to the original implementation.

extrapolate_rate 🔒

Implementations of rate, increase and delta functions in PromQL.

holt_winters 🔒

Implementation of holt_winters in PromQL. Refer to the original implementation.

idelta 🔒

predict_linear 🔒

Implementation of predict_linear in PromQL. Refer to the original implementation.

quantile 🔒

resets 🔒

Implementation of reset in PromQL. Refer to the original implementation.

Structs

AbsentOverTime

absent_over_time returns an empty vector if the range vector passed to it has any elements (floats or native histograms) and a 1-element vector with the value 1 if the range vector passed to it has no elements.

AvgOverTime

The average value of all points in the specified interval.

Changes

used to count the number of value changes that occur within a specific time range

CountOverTime

The count of all values in the specified interval.

Deriv

HoltWinters

There are 3 variants of smoothing functions:

IDelta

The funcIdelta in Promql, from https://github.com/prometheus/prometheus/blob/6bdecf377cea8e856509914f35234e948c4fcb80/promql/functions.go#L235

LastOverTime

The most recent point value in specified interval.

MaxOverTime

The maximum value of all points in the specified interval.

MinOverTime

The minimum value of all points in the specified interval.

PredictLinear

PresentOverTime

the value 1 for any series in the specified interval.

QuantileOverTime

Resets

used to count the number of times the time series starts over.

StddevOverTime

the population standard deviation of the values in the specified interval. Prometheus’s implementation: https://github.com/prometheus/prometheus/blob/f55ab2217984770aa1eecd0f2d5f54580029b1c0/promql/functions.go#L556-L569

StdvarOverTime

the population standard variance of the values in the specified interval. DataFusion’s implementation: https://github.com/apache/arrow-datafusion/blob/292eb954fc0bad3a1febc597233ba26cb60bda3e/datafusion/physical-expr/src/aggregate/variance.rs#L224-#L241

SumOverTime

The sum of all values in the specified interval.

Functions

compensated_sum_inc 🔒

compensation(Kahan) summation algorithm - a technique for reducing the numerical error in floating-point arithmetic. The algorithm also includes the modification (“Neumaier improvement”) that reduces the numerical error further in cases where the numbers being summed have a large difference in magnitude Prometheus’s implementation: https://github.com/prometheus/prometheus/blob/f55ab2217984770aa1eecd0f2d5f54580029b1c0/promql/functions.go#L782

extract_array 🔒

linear_regression 🔒

linear_regression performs a least-square linear regression analysis on the times and values. It return the slope and intercept based on times and values. Prometheus’s implementation: https://github.com/prometheus/prometheus/blob/90b2f7a540b8a70d8d81372e6692dcbb67ccbaaa/promql/functions.go#L793-L837

Type Aliases

Delta

Increase

Rate