promql::range_array

Struct RangeArray

pub struct RangeArray {
    array: DictionaryArray<Int64Type>,
}

An compound logical “array” type. Represent serval ranges (slices) of one array. It’s useful to use case like compute sliding window, or range selector from promql.

It’s build on top of Arrow’s [DictionaryArray]. [DictionaryArray] contains two sub-arrays, one for dictionary key and another for dictionary value. Both of them can be arbitrary types, but here the key array is fixed to u32 type.

            │ ┌─────┬─────┬─────┬─────┐
     Two    │ │ i64 │ i64 │ i64 │ i64 │      Keys
   Arrays   │ └─────┴─────┴─────┴─────┘     (Fixed to i64)
     In     │
  Arrow's   │ ┌────────────────────────────┐
Dictionary  │ │                            │ Values
   Array    │ └────────────────────────────┘(Any Type)

Because the i64 key array is reinterpreted into two u32 for offset and length in RangeArray to represent a “range”:

63            32│31             0
┌───────────────┼───────────────┐
│  offset (u32) │  length (u32) │
└───────────────┼───────────────┘

Then the [DictionaryArray] can be expanded to several ranges like this:

Keys
┌───────┬───────┬───────┐      ┌───────┐
│ (0,2) │ (1,2) │ (2,2) │ ─┐   │[A,B,C]│ values.slice(0,2)
└───────┴───────┴───────┘  │   ├───────┤
 Values                    ├─► │[B,C,D]│ values.slice(1,2)
┌───┬───┬───┬───┬───┐      │   ├───────┤
│ A │ B │ C │ D │ E │     ─┘   │[C,D,E]│ values.slice(2,2)
└───┴───┴───┴───┴───┘          └───────┘

Fields
array: DictionaryArray<Int64Type>
Implementations

impl RangeArray

pub const fn key_type() -> DataType

pub fn value_type(&self) -> DataType

pub fn try_new(dict: DictionaryArray<Int64Type>) -> Result<Self>

pub fn from_ranges<R>(values: ArrayRef, ranges: R) -> Result<Self>
where R: IntoIterator<Item = RangeTuple> + Clone,

pub unsafe fn from_ranges_unchecked<R>(values: ArrayRef, ranges: R) -> Self
where R: IntoIterator<Item = RangeTuple>,

Construct RangeArray from given range without checking its validity.

Safety

Caller should ensure the given range are valid. Otherwise use from_ranges instead.

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn get(&self, index: usize) -> Option<ArrayRef>

pub fn get_offset_length(&self, index: usize) -> Option<(usize, usize)>

pub fn into_dict(self) -> DictionaryArray<Int64Type>

Return the underlying Arrow’s [DictionaryArray]. Notes the dictionary array might be invalid from Arrow’s definition. Be care if try to access this dictionary through Arrow’s API.

fn check_ranges<R>(value_len: usize, ranges: R) -> Result<()>
where R: IntoIterator<Item = RangeTuple>,

pub fn convert_field(field: &Field) -> Field

Change the field’s datatype to the type after processed by RangeArray. Like Utf8 will become Dictionary<Int64, Utf8>.

pub fn convert_data_type(value_type: DataType) -> DataType

Build datatype of wrapped RangeArray on given value type.

pub fn values(&self) -> &ArrayRef

pub fn ranges(&self) -> impl Iterator<Item = Option<RangeTuple>> + '_