promql

Module extension_plan

Modules

empty_metric 🔒

histogram_fold 🔒

instant_manipulate 🔒

normalize 🔒

planner 🔒

range_manipulate 🔒

scalar_calculate 🔒

series_divide 🔒

union_distinct_on 🔒

Structs

EmptyMetric

Empty source plan that generate record batch with two columns:

EmptyMetricExec

EmptyMetricStream

HistogramFold

HistogramFold will fold the conventional (non-native) histogram (1) for later computing.

HistogramFoldExec

HistogramFoldStream

InstantManipulate

Manipulate the input record batch to make it suitable for Instant Operator.

InstantManipulateExec

InstantManipulateStream

PromExtensionPlanner

RangeManipulate

Time series manipulator for range function.

RangeManipulateExec

RangeManipulateStream

ScalarCalculate

ScalarCalculate is the custom logical plan to calculate scalar in PromQL, return NaN when have multiple time series.

SeriesDivide

SeriesDivideExec

SeriesDivideStream

Assume the input stream is ordered on the tag columns.

SeriesNormalize

Normalize the input record batch. Notice that for simplicity, this method assumes the input batch only contains sample points from one time series.

SeriesNormalizeExec

SeriesNormalizeStream

UnionDistinctOn

A special kind of UNION(OR in PromQL) operator, for PromQL specific use case.

UnionDistinctOnExec

UnionDistinctOnStream

Constants

METRIC_NUM_SERIES 🔒

Functions

build_special_time_expr

Type Aliases

Millisecond