Module promql::extension_plan

Modules

empty_metric 🔒

instant_manipulate 🔒

normalize 🔒

planner 🔒

range_manipulate 🔒

series_divide 🔒

Structs

EmptyMetric

EmptyMetricExec

EmptyMetricStream

InstantManipulate

Manipulate the input record batch to make it suitable for Instant Operator.

InstantManipulateExec

InstantManipulateStream

PromExtensionPlanner

RangeManipulate

Time series manipulator for range function.

RangeManipulateExec

RangeManipulateStream

SeriesDivide

SeriesDivideExec

SeriesDivideStream

Assume the input stream is ordered on the tag columns.

SeriesNormalize

Normalize the input record batch. Notice that for simplicity, this method assumes the input batch only contains sample points from one time series.

SeriesNormalizeExec

SeriesNormalizeStream

Type Definitions

Millisecond 🔒