promql::functions

Module aggr_over_time

Structs

AbsentOverTime

absent_over_time returns an empty vector if the range vector passed to it has any elements (floats or native histograms) and a 1-element vector with the value 1 if the range vector passed to it has no elements.

AvgOverTime

The average value of all points in the specified interval.

CountOverTime

The count of all values in the specified interval.

LastOverTime

The most recent point value in specified interval.

MaxOverTime

The maximum value of all points in the specified interval.

MinOverTime

The minimum value of all points in the specified interval.

PresentOverTime

the value 1 for any series in the specified interval.

StddevOverTime

the population standard deviation of the values in the specified interval. Prometheus’s implementation: https://github.com/prometheus/prometheus/blob/f55ab2217984770aa1eecd0f2d5f54580029b1c0/promql/functions.go#L556-L569

StdvarOverTime

the population standard variance of the values in the specified interval. DataFusion’s implementation: https://github.com/apache/arrow-datafusion/blob/292eb954fc0bad3a1febc597233ba26cb60bda3e/datafusion/physical-expr/src/aggregate/variance.rs#L224-#L241

SumOverTime

The sum of all values in the specified interval.

Functions

absent_over_time 🔒

avg_over_time 🔒

count_over_time 🔒

last_over_time 🔒

max_over_time 🔒

min_over_time 🔒

present_over_time 🔒

stddev_over_time 🔒

stdvar_over_time 🔒

sum_over_time 🔒